Auth permission and role create

+++ app/Http/Controllers/Admin/AdminController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Http\Controllers\Services\AuthNeed;
 use App\Models\Admin;
 use App\Models\Role;
 use Illuminate\Http\RedirectResponse;
@@ -12,57 +13,36 @@ use Illuminate\Support\Facades\Hash;
 
 class AdminController extends Controller
 {
-    public function dashboard()
-    {
-        return view('admin.dashboard');
-    }
-    public function showLoginForm()
-    {
-        return view('admin.auth.login');
-    }
 
-    public function login(Request $request)
-    {
-        $credentials = $request->only('email', 'password');
-
-        if (Auth::guard('admin')->attempt($credentials)) {
-            return redirect()->route('admin.dashboard');
-        }
-
-        return back()->withErrors(['email' => 'Invalid credentials.']);
-    }
-    public function destroy(Request $request): RedirectResponse
-    {
-        Auth::guard('admin')->logout();
-
-        return redirect()->route('admin.login');
-    }
     public function create()
     {
         $roles = Role::all();
         return view('admin.users.create', compact('roles'));
     }
+
     public function store(Request $request)
     {
         // Validate form inputs
         $request->validate([
-            'name'      => 'required|string|max:255',
-            'email'     => 'required|email|unique:admins,email',
-            'password'  => 'required|string|min:8',
-            'role_id'   => 'required|exists:roles,id',
+            'name' => 'required|string|max:255',
+            'email' => 'required|email|unique:admins,email',
+            'password' => 'required|string|min:8',
+            'role_id' => 'required|exists:roles,id',
         ]);
 
+        AuthNeed::permission('*')->role(['super_admin', 'admin']);
 
         // Create new admin user
         Admin::create([
-            'name'      => $request->name,
-            'email'     => $request->email,
-            'password'  => Hash::make($request->password),
-            'role_id'   => $request->role_id,
+            'name' => $request->name,
+            'email' => $request->email,
+            'password' => Hash::make($request->password),
+            'role_id' => $request->role_id,
         ]);
 
         return redirect()->route('admin.roles.index')->with('admin', 'New admin user created successfully.');
     }
+
     // Show edit form for admin user
     public function edit($id)
     {
@@ -71,18 +51,29 @@ class AdminController extends Controller
 
         return view('admin.users.edit', compact('user', 'roles'));
     }
-    public function delete($id)
+
+    public function delete($id): RedirectResponse
     {
+        $currentAdmin = Auth::guard('admin')->user();
+
+        AuthNeed::permission('*')->role(['super_admin', 'admin']);
         $user = Admin::findOrFail($id);
-        $user->delete();
+       // dd($user->role->name);
+
+        if ($currentAdmin->id === $user->id || $user->role->name === "super_admin") {
+            return redirect()->back()->with('admin', 'Could not delete this user.');
+        }
+         $user->delete();
         return redirect()->back()->with('admin', 'Admin user deleted successfully.');
+
     }
 
     // Update admin user
     public function update(Request $request, $id)
     {
+        AuthNeed::permission('*')->role(['super_admin', 'admin']);
         $request->validate([
-            'name'  => 'required|string|max:255',
+            'name' => 'required|string|max:255',
             'email' => 'required|email|unique:admins,email,' . $id,
             'role_id' => 'required|exists:roles,id',
         ]);
@@ -92,17 +83,6 @@ class AdminController extends Controller
 
         return redirect()->route('admin.roles.index')->with('admin', 'Admin user updated successfully.');
     }
-    public function changeRole(Request $request, $id)
-    {
-        $request->validate([
-            'role_id' => 'required|exists:roles,id'
-        ]);
 
-        $user = Admin::findOrFail($id);
-        $user->role_id = $request->role_id;
-        $user->save();
-
-        return redirect()->back()->with('role', 'Role updated successfully.');
-    }
 
 }
